rename: inputbox 컴포넌트 세분화를 위한 이름 변경
